Port to TrinityCore 4.3.4 (Cataclysm): compile, boot, and combat AI
Adapt this AzerothCore 3.3.5a playerbot module to ShatterCore (TC 4.3.4):
clean compile/link, stable boot + bot spawning, and reworked combat AI and
talent templates for every class/spec.

API/port: PhaseShift phasing, structured WorldPackets, Item-sparse store
instead of the item_template SQL table, POWER_HOLY_POWER, TOTAL_LOCALES (12),
quest-relation renames; exclude Ai/Raid + Ai/Dungeon (WotLK tactics) from build.

Boot fixes: ValidateAppearance-valid random bot generation, RandomItemMgr null
guards, skip item_template-based caches, PlayerbotTextMgr locale over-read.

Spell learning: InitAvailableSpells enumerates the trainer table directly
(4.3.4 keys trainers by trainer.Id, not creature entry) so bots learn their
trained abilities (seals, Judgement, Exorcism, Avenging Wrath, ...).

Resources: Focus / Soul Shards / Runic Power values (POWER_*), plus per-class
Holy Power, Shadow Orbs (aura stacks), and Eclipse.

Combat AI: rewrote the DPS/Tank/Heal strategies for the 4.3.4 rotations of all
ten classes (Holy-Power Paladin, Focus Hunter, Shadow-Orb Priest, rune DK,
Eclipse Druid, etc.), casting by name with runtime-verified 4.3.4 spell names.

Talents: 31-point level-85 PremadeSpecLink builds for every spec in
conf/playerbots.conf.dist; level cap raised 80 -> 85.

+++ src/Mgr/Item/RandomItemMgr.h
@@ -19,7 +19,7 @@ class ChatHandler;
 
 struct ItemTemplate;
 
-enum EquipmentSlots : uint32;
+enum EquipmentSlots : uint8;
 
 enum RandomItemType
 {
